fixed loop when node fails
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/DJob/trunk@5632 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ DistribJob/lib/perl/Node.pm
@@ -336,16 +336,19 @@ sub cleanUp {
     kill(1, $self->{taskPid}) unless waitpid($self->{taskPid},1);
   }
 
-  print "Cleaning up node $self->{nodeNum}...\n";
+  if($state != $FAILEDNODE){  ## if the node has failed don't want to run commands on it ...
 
-  ##now call the task->cleanUpNode method to enable  users to stop processes running on node
-  my $task = $self->getTask();
-  $task->cleanUpNode($self) if $task;
-
-  if($self->{nodeNum} && $self->getPort()){
-    $self->runCmd("/bin/rm -r $self->{nodeDir}", 1);
-    $self->runCmd("closeAndExit",1);
-    $self->closePort();
+    print "Cleaning up node $self->{nodeNum}...\n";
+    
+    ##now call the task->cleanUpNode method to enable  users to stop processes running on node
+    my $task = $self->getTask();
+    $task->cleanUpNode($self) if $task;
+    
+    if($self->{nodeNum} && $self->getPort()){
+      $self->runCmd("/bin/rm -r $self->{nodeDir}", 1);
+      $self->runCmd("closeAndExit",1);
+      $self->closePort();
+    }
   }
   $self->setState($state ? $state : $COMPLETE); ##complete
 }
